+++ src/app/(home)/components/sections/PopularReviews.tsx
@@ -2,7 +2,7 @@
 
 import { useState, useEffect } from "react";
 import Link from "next/link";
-import Button from "@/components/ui/Buttons/Button";
+import Button from "@/components/common/Buttons/Button";
 import SectionHeader from "../ui/SectionHeader";
 import EmptyState from "../ui/EmptyState";
 import ReviewCard from "../reviews/ReviewCard";
