+++ src/app/(home)/components/ui/FilterButtons.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import RadioButton from "@/components/ui/Buttons/RadioButton";
+import RadioButton from "@/components/common/Buttons/RadioButton";
 
 interface FilterButtonsProps {
   selectedFilter: string;
