+++ src/app/auth/signup/page.tsx
@@ -8,7 +8,7 @@ import { useAuthStore } from "@/store/authStore";
 import Input from "@/components/ui/Input";
 import Link from "next/link";
 import { useRouter } from "next/navigation";
-import Button from "@/components/ui/Buttons/Button";
+import Button from "@/components/common/Buttons/Button";
 
 export default function SignUpPage() {
   const [isPasswordVisible, setIsPasswordVisible] = useState(false);
